#Issue 346 Fixed
These frames are for storing musician credits such as

producer:john
engineer:mick

the trouble is only one such frame is allowed in a tag, and not very to add or 
remove just one musician credit. Furthermore want to add generic keys such as
ENGINEER or ARRANGER but this is difficult for ID3 because these mutiple generic
keys have to not only map to the same type of frame but to the same instance of
the frame.