Support reading of genre codes > 125 in ID3.
Changed behavior so that valid genre codes greater than 125 are translated to strings (e.g. 127 = Drum & Bass) when read.
When writing, don't translate genre strings to codes, if they correspond to codes > 125 (as iTunes won't be able to read them). Rather use the string instead.
However, when setting a string like "(127)" directly, write it just like that (thus allowing the user to force-set the code instead of the string).

Adjusted comparisons with getMaxGenreId() so that they are inclusive (the max id is still valid, so we need to use <=, instead of <).